Created package org.junit.contrib.java.lang.system.internal.
This package stores all classes, which are not part of the public
API of this library.
